Status box: Fix bug from race condition when starting audit
This was only happening in Cypress tests on occassion. When starting an audit, the rounds would load before the jurisdictions would reload with round status, so trying to access it would fail. These data are always loaded at the same time, but not in the same request.

Simple fix is just to check for null - it shouldn't cause real impact.